Rename cypher lib dir (#613)
