+++ timesketch/lib/cypher/insertable_string.py
